requirements.txt
tensorboard
tensorflow-io==0.31.0
boto3
python-dotenv
s3fs